[Refactor]: 기능별로 패키지 구조 변경
