+++ BE/passionroad/src/main/java/com/passionroad/passionroad/SamplerController.java
@@ -1,4 +1,4 @@
-package com.passionroad.passionroad.controller;
+package com.passionroad.passionroad;
 
 import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.GetMapping;
