+++ BE/passionroad/src/main/java/com/passionroad/passionroad/freeboard/domain/search/FreeBoardSearchImpl.java
@@ -1,6 +1,6 @@
-package com.passionroad.passionroad.domain.search;
+package com.passionroad.passionroad.freeboard.domain.search;
 
-import com.passionroad.passionroad.domain.freeboard.FreeBoard;
+import com.passionroad.passionroad.freeboard.domain.freeboard.FreeBoard;
 import com.passionroad.passionroad.domain.freeboard.QFreeBoard;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.JPQLQuery;
