+++ BE/passionroad/src/main/java/com/passionroad/passionroad/freeboard/dto/PageResponseDTO.java
@@ -1,4 +1,4 @@
-package com.passionroad.passionroad.dto;
+package com.passionroad.passionroad.freeboard.dto;
 
 import lombok.Builder;
 import lombok.Getter;
