+++ BE/passionroad/src/main/java/com/passionroad/passionroad/freeboard/repository/FreeBoardRepository.java
@@ -1,7 +1,7 @@
-package com.passionroad.passionroad.repository;
+package com.passionroad.passionroad.freeboard.repository;
 
-import com.passionroad.passionroad.domain.freeboard.FreeBoard;
-import com.passionroad.passionroad.domain.search.FreeBoardSearch;
+import com.passionroad.passionroad.freeboard.domain.freeboard.FreeBoard;
+import com.passionroad.passionroad.freeboard.domain.search.FreeBoardSearch;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
