+++ BE/passionroad/src/main/java/com/passionroad/passionroad/member/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.passionroad.passionroad.domain;
+package com.passionroad.passionroad.member.domain;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
