+++ BE/passionroad/src/main/java/com/passionroad/passionroad/studyroom/controller/StudyRoomController.java
@@ -1,7 +1,7 @@
 package com.passionroad.passionroad.studyroom.controller;
 
 
-import com.passionroad.passionroad.domain.member.Member;
+import com.passionroad.passionroad.member.domain.Member;
 import com.passionroad.passionroad.studyroom.entity.EnterMember;
 import com.passionroad.passionroad.studyroom.entity.MemberDetailsImpl;
 import com.passionroad.passionroad.studyroom.entity.StudyRoom;
