+++ BE/passionroad/src/main/java/com/passionroad/passionroad/studyroom/entity/EnterMember.java
@@ -1,6 +1,6 @@
 package com.passionroad.passionroad.studyroom.entity;
 
-import com.passionroad.passionroad.domain.member.Member;
+import com.passionroad.passionroad.member.domain.Member;
 import com.passionroad.passionroad.studyroom.Timestamped;
 import lombok.Getter;
 import lombok.Setter;
