+++ BE/passionroad/src/main/java/com/passionroad/passionroad/studyroom/entity/StudyRoom.java
@@ -1,6 +1,6 @@
 package com.passionroad.passionroad.studyroom.entity;
 
-import com.passionroad.passionroad.domain.member.Member;
+import com.passionroad.passionroad.member.domain.Member;
 import com.passionroad.passionroad.studyroom.Timestamped;
 import com.passionroad.passionroad.studyroom.request.StudyRoomRequestDto;
 import lombok.*;
