+++ BE/passionroad/src/main/java/com/passionroad/passionroad/studyroom/repository/EnterMemberRepository.java
@@ -1,6 +1,6 @@
 package com.passionroad.passionroad.studyroom.repository;
 
-import com.passionroad.passionroad.domain.member.Member;
+import com.passionroad.passionroad.member.domain.Member;
 import com.passionroad.passionroad.studyroom.entity.EnterMember;
 import com.passionroad.passionroad.studyroom.entity.StudyRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
