+++ BE/passionroad/src/test/java/com/passionroad/passionroad/service/FreeBoardServiceTests.java
@@ -1,10 +1,11 @@
 package com.passionroad.passionroad.service;
 
-import com.passionroad.passionroad.domain.member.Member;
-import com.passionroad.passionroad.dto.FreeBoardDTO;
-import com.passionroad.passionroad.dto.PageRequestDTO;
-import com.passionroad.passionroad.dto.PageResponseDTO;
-import com.passionroad.passionroad.repository.MemberRepository;
+import com.passionroad.passionroad.freeboard.service.FreeBoardService;
+import com.passionroad.passionroad.member.domain.Member;
+import com.passionroad.passionroad.freeboard.dto.FreeBoardDTO;
+import com.passionroad.passionroad.freeboard.dto.PageRequestDTO;
+import com.passionroad.passionroad.freeboard.dto.PageResponseDTO;
+import com.passionroad.passionroad.member.repository.MemberRepository;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
